+++ src/main/kotlin/dev/turingcomplete/intellijdevelopertoolsplugin/_internal/tool/editor/PsiJavaUtils.kt
@@ -21,35 +21,42 @@ internal object PsiJavaUtils {
   // -- Initialization ---------------------------------------------------------------------------------------------- //
   // -- Exported Methods -------------------------------------------------------------------------------------------- //
 
-  fun getTextFromStringValueOrIdentifier(e: AnActionEvent): Pair<String, TextRange>? {
+  fun getPsiElementAtCaret(e: AnActionEvent): PsiElement? {
     val psiFile = e.getData(CommonDataKeys.PSI_FILE) ?: return null
     val editor = e.getData(CommonDataKeys.EDITOR) ?: return null
     if (editor.getSelectedText() != null) {
       return null
     }
 
-    val psiElement = psiFile.findElementAt(editor.caretModel.offset) ?: return null
-    return getTextIfStringValueOrIdentifier(psiElement)
+    return psiFile.findElementAt(editor.caretModel.offset)
   }
 
   fun getTextIfStringValueOrIdentifier(psiElement: PsiElement): Pair<String, TextRange>? {
-    return if (psiElement is PsiJavaToken && psiElement.elementType == JavaTokenType.STRING_LITERAL) {
+    return getTextIfStringValue(psiElement)
+      ?: if (psiElement is PsiIdentifier) {
+        psiElement.text to psiElement.textRange
+      }
+      else {
+        null
+      }
+  }
+
+  fun getTextIfStringValue(psiElement: PsiElement): Pair<String, TextRange>? =
+    if (psiElement is PsiJavaToken && psiElement.elementType == JavaTokenType.STRING_LITERAL) {
       // Remove the enclosing quotations
       val newStart = psiElement.textRange.start + 1
       val newEnd = psiElement.textRange.end - 1
       if (newStart > newEnd) {
-        return null
+        null
       }
-      psiElement.text.substring(1, psiElement.text.length - 1) to TextRange(newStart, newEnd)
-    }
-    else if (psiElement is PsiIdentifier) {
-      psiElement.text to psiElement.textRange
+      else {
+        psiElement.text.substring(1, psiElement.text.length - 1) to TextRange(newStart, newEnd)
+      }
     }
     else {
       null
     }
-  }
-  
+
   // -- Private Methods --------------------------------------------------------------------------------------------- //
   // -- Inner Type -------------------------------------------------------------------------------------------------- //
 }